requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai

google-generativeai

streamlit

numpy
pandas
scikit-learn
joblib

python-dotenv
requests
pydantic

pytest
